+++ mayan/apps/permissions/icons.py
@@ -6,7 +6,8 @@ icon_role_create = Icon(
     driver_name='fontawesome-dual', primary_symbol='user-secret',
     secondary_symbol='plus'
 )
-icon_role_delete = Icon(driver_name='fontawesome', symbol='times')
+icon_role_delete_single = Icon(driver_name='fontawesome', symbol='times')
+icon_role_delete_multiple = icon_role_delete_single
 icon_role_edit = Icon(driver_name='fontawesome', symbol='pencil-alt')
 icon_role_groups = Icon(driver_name='fontawesome', symbol='users')
 icon_role_list = Icon(driver_name='fontawesome', symbol='user-secret')